[IMP] fieldservice_sale: Remove the recurring part
