+++ fieldservice_sale/models/__init__.py
@@ -1,8 +1,10 @@
-# Copyright (C) 2018 - TODAY, Open Source Integrators
+# Copyright (C) 2018 Open Source Integrators
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import fsm_location
-from . import product
-from . import sale_order
-from . import fsm_order
-from . import fsm_recurring
+from . import (
+    fsm_location,
+    product_template,
+    sale_order,
+    sale_order_line,
+    fsm_order
+)
