--- fieldservice_sale/models/fsm_recurring.py
@@ -1,21 +1,0 @@
-# Copyright (C) 2019 - TODAY, Brian McMaster, Open Source Integrators
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from odoo import fields, models, _
-
-
-class FSMRecurring(models.Model):
-    _inherit = 'fsm.recurring'
-
-    sale_line_id = fields.Many2one('sale.order.line')
-
-    def action_view_sales(self):
-        self.ensure_one()
-        return {
-            "type": "ir.actions.act_window",
-            "res_model": "sale.order",
-            "views": [[False, "form"]],
-            "res_id": self.sale_line_id.order_id.id,
-            "context": {"create": False},
-            "name": _("Sales Orders"),
-        }
