+++ fieldservice_sale/readme/CONTRIBUTORS.rst
@@ -1,6 +1,9 @@
-* Steve Campbell <scampbell@opensourceintegrators.com>
-* Maxime Chambreuil <mchambreuil@opensourceintegrators.com>
-* Wolfgang Hall <whall@opensourceintegrators.com>
+* Open Source Integrators <https://opensourceintegrators.com>
+
+  * Steve Campbell <scampbell@opensourceintegrators.com>
+  * Maxime Chambreuil <mchambreuil@opensourceintegrators.com>
+  * Wolfgang Hall <whall@opensourceintegrators.com>
+
 * Serpent Consulting Services Pvt. Ltd. <support@serpentcs.com>
 * Brian McMaster <brian@mcmpest.com>
 * Raphaël Reverdy <raphael.reverdy@akretion.com>
