--- fieldservice_sale/readme/CREDITS.rst
@@ -1,3 +1,0 @@
-The development of this module has been financially supported by:
-
-* Open Source Integrators <https://opensourceintegrators.com>
