+++ fieldservice_sale/readme/DESCRIPTION.rst
@@ -1,1 +1,2 @@
-The module integrate the field service application with the sales one.
+The module integrates the Field Service application with the Sales one and
+allows you to sell field service products.
