+++ fieldservice_sale/views/fsm_location.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
     <!-- Field Service Location Form View-->
